code/requirements.txt
Flask
datetime
shutil
python-magic